requirements.txt
pyTelegramBotAPI==4.15.4
apscheduler
# Для экспорта в Google Sheets:
gspread
google-auth 